status update done

+++ lib/actions/userEvent.actions.ts
@@ -23,7 +23,7 @@ import { handleError } from "../utils";
 export const createUserEvent = async (
   userId: string,
   eventId: string,
-  token: string // Expect the JWT token to be passed
+  token: string
 ): Promise<IResponse<IUserEvent | string | jwt.JwtPayload>> => {
   const tokenResponse = await verifyToken(token, [UserRole.Admin]);
   if (tokenResponse.status === ResponseStatus.Error) {
@@ -35,12 +35,11 @@ export const createUserEvent = async (
 
     await connectToDatabase();
 
-    // Check if there is already a userEvent with the same userId and eventId
+    // Check for existing relationship
     const existingUserEvent = await UserEvent.findOne({
       user: userId,
       event: eventId,
     });
-
     if (existingUserEvent) {
       return {
         status: ResponseStatus.Error,
@@ -49,9 +48,8 @@ export const createUserEvent = async (
       };
     }
 
-    // Create a new userEvent if it doesn't exist
+    // Create a new UserEvent
     const userEvent = await UserEvent.create({ user: userId, event: eventId });
-
     return {
       status: ResponseStatus.Success,
       message: "User-event relationship created successfully",
@@ -59,26 +57,23 @@ export const createUserEvent = async (
       field: JSON.parse(JSON.stringify(userEvent)),
     };
   } catch (error) {
-    const serverError = handleServerError(error);
-    if (serverError) return serverError;
-    return handleError(error);
+    return handleServerError(error) || handleError(error);
   }
 };
 
 // Get user events by event ID (protected)
 export const getUserEventsByEventId = async (
   eventId: string,
-  token: string // Expect the JWT token to be passed
+  token: string
 ): Promise<IResponse<IUserEvent[] | string | jwt.JwtPayload>> => {
   const tokenResponse = await verifyToken(token, [UserRole.User]);
   if (tokenResponse.status === ResponseStatus.Error) {
-    return tokenResponse; // Return unauthorized if token is invalid
+    return tokenResponse;
   }
 
   try {
     await connectToDatabase();
 
-    // Fetch user events by event ID
     const userEvents: IUserEvent[] = await UserEvent.find({
       event: eventId,
     }).populate({ path: "user", select: "-password", model: User });
@@ -90,9 +85,7 @@ export const getUserEventsByEventId = async (
       field: JSON.parse(JSON.stringify(userEvents)),
     };
   } catch (error) {
-    const serverError = handleServerError(error);
-    if (serverError) return serverError;
-    return handleError(error);
+    return handleServerError(error) || handleError(error);
   }
 };
 
@@ -100,23 +93,17 @@ export const getUserEventsByEventId = async (
 export const removeUserEvent = async (
   userId: string,
   eventId: string,
-  token: string // Expect the JWT token to be passed
+  token: string
 ): Promise<IResponse<string | jwt.JwtPayload>> => {
   const tokenResponse = await verifyToken(token, [UserRole.Admin]);
   if (tokenResponse.status === ResponseStatus.Error) {
-    return tokenResponse; // Return unauthorized if token is invalid
+    return tokenResponse;
   }
 
   try {
     await connectToDatabase();
 
-    // Find and remove all user events based on user ID and event ID
-    const result = await UserEvent.deleteMany({
-      user: userId,
-      event: eventId,
-    });
-
-    // Check if any user events were deleted
+    const result = await UserEvent.deleteMany({ user: userId, event: eventId });
     if (result.deletedCount === 0) {
       return {
         status: ResponseStatus.Error,
@@ -132,40 +119,37 @@ export const removeUserEvent = async (
       field: "Success",
     };
   } catch (error) {
-    const serverError = handleServerError(error);
-    if (serverError) return serverError;
-    return handleError(error);
+    return handleServerError(error) || handleError(error);
   }
 };
 
 // Get user events (protected) with pagination
 export const getUserEvents = async (
   userId: string,
-  token: string, // Expect the JWT token to be passed
-  page: number = 1, // Default to the first page
-  limit: number = 10 // Default limit for number of user events per page
+  token: string,
+  page: number = 1,
+  limit: number = 10
 ): Promise<IResponse<IUserEvent[] | string | jwt.JwtPayload>> => {
   const tokenResponse = await verifyToken(token, [
     UserRole.User,
     UserRole.Admin,
   ]);
   if (tokenResponse.status === ResponseStatus.Error) {
-    return tokenResponse; // Return unauthorized if token is invalid
+    return tokenResponse;
   }
 
   try {
     await connectToDatabase();
 
     const skip = (page - 1) * limit;
-
     const userEvents: IUserEvent[] = await UserEvent.find({ user: userId })
       .skip(skip)
       .limit(limit)
       .populate({
         path: "event",
         populate: [
-          { path: "organizer", select: "-password" }, // Fetch organizer details without password
-          { path: "category" }, // Fetch category details
+          { path: "organizer", select: "-password" },
+          { path: "category" },
         ],
       });
 
@@ -181,9 +165,7 @@ export const getUserEvents = async (
       currentPage: page,
     };
   } catch (error) {
-    const serverError = handleServerError(error);
-    if (serverError) return serverError;
-    return handleError(error);
+    return handleServerError(error) || handleError(error);
   }
 };
 
@@ -222,34 +204,6 @@ export const getEventUsers = async (
       totalCount: totalEventUsers,
       totalPages: Math.ceil(totalEventUsers / limit),
       currentPage: page,
-    };
-  } catch (error) {
-    const serverError = handleServerError(error);
-    if (serverError) return serverError;
-    return handleError(error);
-  }
-};
-
-// Update user event status (protected)
-export const updateUserEventStatus = async (
-  userEventId: string,
-  newStatus: UserEventStatus,
-  token: string // Expect the JWT token to be passed
-): Promise<IResponse<string | jwt.JwtPayload>> => {
-  const tokenResponse = await verifyToken(token, [UserRole.Admin]);
-  if (tokenResponse.status === ResponseStatus.Error) {
-    return tokenResponse; // Return unauthorized if token is invalid
-  }
-
-  try {
-    await connectToDatabase();
-    await UserEvent.findByIdAndUpdate(userEventId, { status: newStatus });
-
-    return {
-      status: ResponseStatus.Success,
-      message: "User event status updated successfully",
-      code: 200,
-      field: "Success",
     };
   } catch (error) {
     const serverError = handleServerError(error);
@@ -319,8 +273,6 @@ export const getUserEventsByCategoryId = async (
   page: number = 1,
   limit: number = 10
 ): Promise<IResponse<IUserEvent[] | string | jwt.JwtPayload>> => {
-  console.log(token);
-
   const tokenResponse = await verifyToken(token, [
     UserRole.User,
     UserRole.Admin,
@@ -353,8 +305,6 @@ export const getUserEventsByCategoryId = async (
       .skip((page - 1) * limit)
       .limit(limit);
 
-    console.log(userEvents);
-
     if (!userEvents.length) {
       return {
         status: ResponseStatus.Success,
@@ -383,5 +333,36 @@ export const getUserEventsByCategoryId = async (
     const serverError = handleServerError(error);
     if (serverError) return serverError;
     return handleError(error);
+  }
+};
+
+// Update user event status (protected)
+export const updateUserEventStatus = async (
+  userEventId: string,
+  newStatus: UserEventStatus,
+  token: string // Expect the JWT token to be passed
+): Promise<IResponse<string | jwt.JwtPayload>> => {
+  const tokenResponse = await verifyToken(token, [UserRole.User]);
+  if (tokenResponse.status === ResponseStatus.Error) {
+    return tokenResponse; // Return unauthorized if token is invalid
+  }
+
+  try {
+    await connectToDatabase();
+    const update = await UserEvent.findByIdAndUpdate(userEventId, {
+      status: newStatus,
+    });
+    console.log(update);
+
+    return {
+      status: ResponseStatus.Success,
+      message: "User event status updated successfully",
+      code: 200,
+      field: "Success",
+    };
+  } catch (error) {
+    console.log(error);
+
+    return handleError(error) || handleServerError(error);
   }
 };
